pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "l3-data-agent"
version = "0.1.0"
description = "L3 Autonomous Data Agent — playbook-driven data analysis with LLM"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "litellm>=1.40.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio"]
mysql = ["pymysql>=1.1.0"]
postgresql = ["psycopg2-binary>=2.9.0"]
duckdb = ["duckdb-engine>=0.9.0"]
all = ["pymysql>=1.1.0", "psycopg2-binary>=2.9.0", "duckdb-engine>=0.9.0"]

[project.scripts]
l3-agent = "l3_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
